LIBITD-2328. Added "docs/ConformanceTests.md" document
Added "docs/ConformanceTests.md" describing end-to-end tests that
actually attempt uploads to the AWS bucket, to verify that the
"deposit" command correctly handles both success and failure cases.

In the "README.md" file, added a link to "docs/ConformanceTests.md" and
added language specifications to code blocks in accordance with SSDR
policy.

https://umd-dit.atlassian.net/browse/LIBITD-2328